fixed backend functions and added a few helper functions: grabbing data from data table now works, removing data now works, assigning data to users now works

+++ Computer_App/db.py
@@ -1,14 +1,15 @@
 import mysql.connector
-import difflib
+
+userDict = {}
+
 #initalizes database if table doesn't already exist
-def addToUserTable(user, pw):
+def addToUserTable(user, pw, fname, lname):
     (conn, cursor) = createCon('cs115','insecuurity')
     print(user, pw)
     try:
         create = ("""CREATE USER %s IDENTIFIED BY %s""")
         cursor.execute(create, user, pw)
-        conn.commit()
-        permissions = ("""GRANT SELECT,DELETE,INSERT,UPDATE on data to %s identified by %s""")
+        permissions = ("""GRANT SELECT,DELETE,INSERT,UPDATE on secuure.data to %s identified by %s""")
         cursor.execute(permissions, user, pw)
         conn.commit()
     except mysql.connector.Error as e:
@@ -21,6 +22,7 @@ def addToUserTable(user, pw):
             first_name  text,
             last_name  text
             )""")
+    insertToUserTable(fname, lname, user)
     conn.close()
 
 
@@ -33,17 +35,16 @@ def getUserIdNum():
         if id[0] > max_id:
             max_id = id[0]
     conn.close()
-    return max_id + 1
+    return max_id
 
-def getDataIdNum():
+#gets highest ID number in the DATA table -- need two different functions because user permissions will be different
+def getUserIdForData(username):
     (conn, cursor) = createCon('cs115','insecuurity')
-    cursor.execute("""SELECT userid from data""")
-    max_id = 0
-    for id in cursor:
-        if id[0] > max_id:
-            max_id = id[0]
-    conn.close()
-    return max_id + 1
+    select_uid = """SELECT id, username from users where username='%s' """ %username
+    cursor.execute(select_uid)
+    for i, u in cursor:
+        return i
+
 
 def createPassTable():
     (conn, cursor) = createCon('cs115','insecuurity')
@@ -55,19 +56,24 @@ def createPassTable():
                 notes text
                 )""")
     conn.close()
-#inserts into the "accounts" table.
+
+#inserts into the "users" table.
 #Thing to note: table to be inserted into CANNOT be a variable (must be hardcoded)
 def insertToUserTable(fname,lname, user):
     (conn, cursor) = createCon('cs115','insecuurity')
+    max_id = getUserIdNum()
+
+    print(max_id)
+    if (user, max_id) in userDict.items() is None:
+        userDict[user] = max_id # assigns a global table of matching id names to account names
+
     query = ("""SELECT username FROM users""")       #
     cursor.execute(query)                           #
     for u in cursor:                                # This piece checks if an account exists already
-        print(u[0], user)
         if u[0].lower() == user.lower():           #
              print("Account name already exists")    #
              return                                  #
-    max_id = getUserIdNum()
-    cursor.execute("""INSERT IGNORE INTO users values (%s, %s, %s, %s)""", (max_id, user, fname, lname))
+    cursor.execute("""INSERT IGNORE INTO users values (%s, %s, %s, %s)""", (max_id+1, user, fname, lname))
     conn.commit()
     conn.close()
 
@@ -111,31 +117,29 @@ def addPassForWebsite(username, pw, website, notes):
             print("Duplicate entry in table, please try again")
             conn.close
             return
-    cursor.execute("""INSERT IGNORE INTO data values (%s, %s, %s, %s, %s)""", (getDataIdNum(), username, website, pw, notes))
+    cursor.execute("""INSERT IGNORE INTO data values (%s, %s, %s, %s, %s)""", (getUserIdForData(username), username, website, pw, notes))
     conn.commit()
     conn.close()
 
 #Prints passwords for a specified user
 def getPasswordsForUser(accountName):
     (conn, cursor) = createCon('cs115','insecuurity')
-    query = ("""SELECT account, password, website, notes FROM DATA """)
+
+    query=("""SELECT account, password, website, notes FROM DATA WHERE userid=%s""" %getUserIdForData(accountName))
     cursor.execute(query)
-    data = []
     for a, p, w, n in cursor:
         if accountName.lower() == a.lower():
             print(a, p, w, n)
     conn.close()
-    return data
 
 
 #Removes entry from the table, all values have to match
 def removeEntry(username, pw, website, notes):
     (conn, cursor) = createCon('cs115', 'insecuurity')
-    query = """DELETE FROM DATA WHERE
-                account=%s && website=%s && password=%s && notes=%s
-                """
+    query = """DELETE FROM DATA WHERE userid=%s && account='%s' && website='%s' && password='%s' && notes='%s'""" %(getUserIdForData(username), username, website, pw, notes, )
 
-    cursor.execute(query, (username, website, pw, notes))
+    print (query)
+    cursor.execute(query)
     conn.commit()
     conn.close()
 
@@ -144,15 +148,14 @@ def removeEntry(username, pw, website, notes):
 #      Testing      #
 #####################
 
-addToUserTable("josccking","root")
-#createPassTable()
-insertToUserTable('John', 'King', 'joscddkingking')
-#print("Before printing\n")
-addPassForWebsite("sup yossssssss", "mypass3!!!21test", "gmail", "last")
-addPassForWebsite("sup yos", "mypass321test", "gmail", "last")
+addToUserTable('jking','root', 'John', 'King')
+getUserIdForData('jking')
+print("Before printing\n")
+addPassForWebsite("jking", "mypass3!!!21test", "gmail", "last")
+addPassForWebsite("jking", "mypass321test", "yahoo", "last")
 getPasswordsForUser("jking")
 print("After printing\n")
-#removeEntry("sup yos", "mypass321test", "gmail", "last")
+removeEntry("jking", "mypass3!!!21test", "gmail", "last")
 getPasswordsForUser("jking")
